Bug 33237: Add display of which language default language is expected
This patch adds the display of 'Default language' to the 'Default'
language in the notices editor tool.

This is so that librarians know which language they are expected to be
writing the notice in so we can remain consistent in both template and
include language used.

Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>
Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>
(cherry picked from commit dc554790439658e5a7b5e1bda62a8d0aa9e2dc3a)
Signed-off-by: Fridolin Somers <fridolin.somers@biblibre.com>