Bug 36945: Fix several missed instance of breadcrumb WRAPPER use
This patch corrects some templates which never got the breadcrumb
WRAPPER update. Some breadcrumbs and page titles have been updated to be
more consistent with page headings.

To test, apply the patch and enable the "CatalogConcerns" system
preference if necessary. Check the following pages to confirm that the
breadcrumbs menu looks correct and that breadcrumbs are consistent with
page titles and page headings.

- Circulation -> Upload offline circulation file (.koc)
  - Upload file -> Add to offline circulation queue
  - View pending offline circulation actions
- Cataloging -> Cataloging concerns
- Reports -> Cash register stats

Sponsored-By: Athens County Public Libraries
Signed-off-by: David Nind <david@davidnind.com>
Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>